Feat DTD parser
Add DTD parser class
Functionality of the class:
Given a file name open the file and load the DTD content form it.
Once loaded as a string the content of the file is parsed into tokens.
Tokens represent some text between < and >, aka tags.
After the content is split into tokens they are parsed into Elements
and Attributes. Elements are stored as a dictionary, where the key
is the element name and the value is a list of the element's children.
The attributes are stored as a dictionary, where the key is the
element name to which the attribute belongs to and the value is
an DTDAttribute object which contains all the specifics of the attribute.