Fix XMLDocument to better work with the generator
Add methods to look for the last element by name
Switch the default function for finding element to add
new sub element, attribute, content to find_last, not
find_first as it was until now. This works better
for the generator and makes more sense to edit the last
element added, not the first.