Fix DTD Attribute class: remove enumerated value
Remove default enumerated value from DTDAttribute.
The field self.value is now used to store the default
enumerated value. default_enumerated_value was used
only by the dtd_parser. It's now changed to properly
use the field value instead.
DTDElement debug print is changed to better print the state
of the object. Now it prints the number of sub elements,
which is more relevant when debugging.